% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols.R
\name{add_cols}
\alias{add_cols}
\title{Add columns to ensure compatibility with previous EM dataset}
\usage{
add_cols(df, cols = c("mesh", "quality", "camera"))
}
\arguments{
\item{df}{A data frame. Usually the output of BBimport("path_to_files")}

\item{cols}{A data frame. Usually the output of fix.CQ("path_to_bycatch_data")}
}
\value{
data.frame object
}
\description{
Add columns to ensure compatibility with previous EM dataset
}
