% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_variables.R
\name{add_variables}
\alias{add_variables}
\title{Add variables to the extracted EM dataset
Take in the BlackBox data from the BBimport function and adds a number of important parameters}
\usage{
add_variables(x = data_work, give_me_more = give_me_more)
}
\arguments{
\item{x}{Data output from function BBimport}

\item{give_me_more}{Should optional variables not relevant for RDBES be created too (default is TRUE)?}
}
\value{
Same dataste with additional columns
}
\description{
Add variables to the extracted EM dataset
Take in the BlackBox data from the BBimport function and adds a number of important parameters
}
