% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFLALO_import.R
\name{EFLALO_import}
\alias{EFLALO_import}
\title{Format logbook / landings data to merge with EM data
Dataset preparations
Format the input dataset following the EFLALO2 way (google TACSAT and EFLALO Formats and check the Github page)}
\usage{
EFLALO_import(x, study_period = NULL)
}
\arguments{
\item{x}{path to the directory where the logbook & sales notes are stored as .csv}

\item{study_period}{A vector of years - e.g., c(2010:2020) - default is NULL}
}
\value{
A dataset with all notes/annotations in long format, where rows are unique for hauls for no or one bycatch within that haul (each additional bycatch is listed as one supplementary row).
}
\description{
Format logbook / landings data to merge with EM data
Dataset preparations
Format the input dataset following the EFLALO2 way (google TACSAT and EFLALO Formats and check the Github page)
}
