% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.CQ.R
\name{fix.CQ}
\alias{fix.CQ}
\title{Fix the catch quantification export format}
\usage{
fix.CQ(
  x =
    "Q:/10-forskningsprojekter/faste-cctv-monitoring/data/blackbox extractions/catch_quantification/",
  incl.fish = FALSE
)
}
\arguments{
\item{x}{Path to the catch quantification files sorted by year or by vessel}

\item{incl.fish}{Should the bycatch data also include (targeted) fish species (e.g. lumpfish, mackerel, etc.)? Defaults to FALSE}
}
\value{
a dataset
}
\description{
Fix the catch quantification export format
}
