% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.hf.R
\name{gcd.hf}
\alias{gcd.hf}
\title{Radial distance between spatial points}
\usage{
gcd.hf(long1, lat1, long2, lat2)
}
\arguments{
\item{long1}{longitude start (decimal)}

\item{lat1}{latitude start (decimal)}

\item{long2}{longitude end (decimal)}

\item{lat2}{latitude end (decimal)}
}
\value{
distance in metres (as.numeric)
}
\description{
Radial distance between spatial points
}
