% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.depth.R
\name{get.depth}
\alias{get.depth}
\title{Extract depth from pre-made raster file
This scripts extracts the depth (metres) for each point in a dataset.
It also adds the ICES subrectangle these points are in.}
\usage{
get.depth(
  x,
  path.to.raster = "Q:/scientific-projects/cctv-monitoring/data/GIS/D5_2020.tif"
)
}
\arguments{
\item{x}{dataset with geographic coordinates in decimal as lon/lat}

\item{path.to.raster}{Path to the a raster file with depth info}
}
\value{
Same dataset with added variables
}
\description{
Extract depth from pre-made raster file
This scripts extracts the depth (metres) for each point in a dataset.
It also adds the ICES subrectangle these points are in.
}
