% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.d2shore.R
\name{get_d2shore}
\alias{get_d2shore}
\title{Calculates the distance to the nearest point on shore
This scripts estimates the distance to shore (d2shore) in metres for each row in the input dataset, assuming ESPG:4326. Missing values are NA_numeric_
If no shapefile is input, Europe is used as default (with ESPG:3035)}
\usage{
get_d2shore(x = logbook, shapefile = NULL, crs_src = 4326)
}
\arguments{
\item{x}{data.frame or data.table object with coordinates as 'LE_LON' and 'LE_LAT' in decimal}

\item{shapefile}{ShapeFile to use - Europe as default}

\item{crs_src}{CRS of the points in the dataset (deafaults to 4326)}
}
\value{
A vector (numeric) of depth values in metre
}
\description{
Calculates the distance to the nearest point on shore
This scripts estimates the distance to shore (d2shore) in metres for each row in the input dataset, assuming ESPG:4326. Missing values are NA_numeric_
If no shapefile is input, Europe is used as default (with ESPG:3035)
}
