% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_depth_EMODNET.R
\name{get_depth_EMODNET}
\alias{get_depth_EMODNET}
\title{Obtain the depth at a point (lat/lon) from EMODnet Digital Bathymetry (DTM 2024) and using RERDDAP
This scripts extracts the depth (metres) for each point in a dataset. Missing values are NA_numeric_}
\usage{
get_depth_EMODNET(lat, lon)
}
\arguments{
\item{lat}{decimal latitude}

\item{lon}{decimal longitude}
}
\value{
A vector (numeric) of depth values in metre
}
\description{
Obtain the depth at a point (lat/lon) from EMODnet Digital Bathymetry (DTM 2024) and using RERDDAP
This scripts extracts the depth (metres) for each point in a dataset. Missing values are NA_numeric_
}
