% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.mode.R
\name{get_mode}
\alias{get_mode}
\title{Replace NA values in a variable with the most frequent (mode) value of that
variable}
\usage{
get_mode(x)
}
\arguments{
\item{x}{The variable for which to get the mode}
}
\value{
Same dataset with updated variable x
}
\description{
Replace NA values in a variable with the most frequent (mode) value of that
variable
}
