% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ices.subrect.R
\name{ices.subrect}
\alias{ices.subrect}
\title{Figures out ICES subrectangle from lon/lat
I stole this script from someone on the internet. Thank you stranger for your unknown contribution to my work.}
\usage{
ices.subrect(lon, lat)
}
\arguments{
\item{lon}{longitude (in decimal)}

\item{lat}{latitude (in decimal)}
}
\value{
Same dataset with added variables
}
\description{
Figures out ICES subrectangle from lon/lat
I stole this script from someone on the internet. Thank you stranger for your unknown contribution to my work.
}
