% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Merge a list of similar csv files in a directory
Read, format, and merge Notes and Annotations from Black Box Analyzer}
\usage{
load_data(x)
}
\arguments{
\item{x}{path to the directory where the logbook & sales notes are stored as .csv}
}
\value{
A list of data frames
}
\description{
Merge a list of similar csv files in a directory
Read, format, and merge Notes and Annotations from Black Box Analyzer
}
