% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_notes_and_CQ.R
\name{merge_notes_and_CQ}
\alias{merge_notes_and_CQ}
\title{Add bycatch records (and catch records if incl.fish = TRUE) from the catch
quantification dataset in Analyzer to the Notes and Annotations dataset}
\usage{
merge_notes_and_CQ(
  x = NotesAnnotations,
  y = CatchQuantification,
  spp_list = list(),
  incl.fish = FALSE,
  rm_errors = TRUE
)
}
\arguments{
\item{x}{A data frame. Usually the output of BBimport("path_to_files")}

\item{y}{A data frame. Usually the output of fix.CQ("path_to_bycatch_data")}

\item{spp_list}{A species list must be provided. The format is an R list of character vectors. The names must correspond to the ones used in BlackBox Analyzer Catch Quantification records.}

\item{incl.fish}{Include info on fish catches (defaults to FALSE)}

\item{rm_errors}{Defaults to TRUE. Removes the problematic bycatch events (those to fix manually) and prints them}
}
\value{
data.frame object
}
\description{
Add bycatch records (and catch records if incl.fish = TRUE) from the catch
quantification dataset in Analyzer to the Notes and Annotations dataset
}
