% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp.in.CQ.R
\name{spp.in.CQ}
\alias{spp.in.CQ}
\title{Which species are marked at least one in the Catch Quantification tables?}
\usage{
spp.in.CQ(
  x =
    "Q:/10-forskningsprojekter/faste-cctv-monitoring/data/blackbox extractions/catch_quantification/"
)
}
\arguments{
\item{x}{Path to the catch quantification files sorted by year or by vessel}
}
\value{
a vector of unique species code names in the CQ tables
}
\description{
Which species are marked at least one in the Catch Quantification tables?
}
