% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.list.R
\name{spp.list}
\alias{spp.list}
\title{Load a species list - Experimental}
\usage{
spp.list(path_to_spp_lists = path_to_spp_lists)
}
\arguments{
\item{path_to_spp_lists}{Path to the a directory where csv files with the list of bycatch species is kept}
}
\value{
list of vectors of species groups
}
\description{
Load a species list - Experimental
}
\examples{
# This function is called in add_bycatch_records and is not supposed to be used on its own.
# For testing purpose: Create bycatch species "lists" (as in ICES spp list, not R list!)
using default for bycatch monitoring in DK 2024
spp.list <- spp.list(path_to_spp_lists =
  "Q:/10-forskningsprojekter/faste-cctv-monitoring/data/species lists/")
}
