% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp.list.table.R
\name{spp.list.table}
\alias{spp.list.table}
\title{Species list to data frame with sc. names and AphiaID}
\usage{
spp.list.table(x = unique.spp)
}
\arguments{
\item{x}{Unique species codes from CQ (the output of ggleR::spp.in.CQ(path_to_bycatch_data))}
}
\value{
a dataset
}
\description{
Species list to data frame with sc. names and AphiaID
}
