% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_variables.R
\name{add_variables}
\alias{add_variables}
\title{Add variables to the extracted EM dataset
Take in the Black Box data from the BBimport function and adds a number of important parameters}
\usage{
add_variables(
  x = data_work,
  study_period = NULL,
  give_me_more = TRUE,
  path_to_soak =
    "Q:/10-forskningsprojekter/faste-cctv-monitoring/data/blackbox extractions/soak/",
  path.to.coastline =
    "Q:/20-forskning/12-gis/Dynamisk/GEOdata2020/BasicLayers/Coastlines/Europe/EEA Europe/EEA_Coastline_20170228.shp",
  path.to.raster = "Q:/10-forskningsprojekter/faste-cctv-monitoring/data/GIS/D5_2020.tif",
  path.to.old.notes =
    "Q:/10-forskningsprojekter/faste-cctv-monitoring/data/processed data/NotesAnnotations.csv"
)
}
\arguments{
\item{x}{Data output from function BBimport}

\item{study_period}{Should all years (default), or only specific years be compiled (defined as c(year1, year2, etc.))?}

\item{give_me_more}{Should optional variables not relevant for RDBES be created too (default is TRUE)?}

\item{path_to_soak}{Some soak time info is missing (not found in BB). Manual fix based on a prior extraction (input is a csv file)}

\item{path.to.coastline}{ESRI Shapefile of the region (used to estimate distance to shore). By default, the file is Europe (Bounding box:  xmin: 943609.8 ymin: -375446 xmax: 7601958 ymax: 6825119; Projected CRS: ETRS89-extended / LAEA Europe)}

\item{path.to.raster}{Raster of depth}

\item{path.to.old.notes}{CSV file wit h a previous version of Notes and Annotations}
}
\value{
Same dataset with additional columns
}
\description{
Add variables to the extracted EM dataset
Take in the Black Box data from the BBimport function and adds a number of important parameters
}
