% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBimport.R
\name{BBimport}
\alias{BBimport}
\title{Extracts Notes and Annotations
Read, format, and merge Notes and Annotations from Black Box Analyzer}
\usage{
BBimport(
  x =
    "Q:/10-forskningsprojekter/faste-cctv-monitoring/data/blackbox extractions/annotations_notes/",
  by.year = TRUE,
  incl.fish = FALSE
)
}
\arguments{
\item{x}{EM data annotations/notes with geographic coordinates in decimal as lon/lat}

\item{by.year}{Are the files sorted by year (default)?}

\item{incl.fish}{Should the bycatch data also include (targeted) fish species (e.g. lumpfish, mackerel, etc.)? Defaults to FALSE}
}
\value{
A dataset with all notes/annotations in long format, where rows are unique for hauls for no or one bycatch within that haul (each additional bycatch is listed as one supplementary row).
}
\description{
Extracts Notes and Annotations
Read, format, and merge Notes and Annotations from Black Box Analyzer
}
