% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.CQ.R
\name{fix.CQ}
\alias{fix.CQ}
\title{Fix the catch quantification export format}
\usage{
fix.CQ(
  x =
    "Q:/10-forskningsprojekter/faste-cctv-monitoring/data/blackbox extractions/catch_quantification/",
  spp_list = NULL
)
}
\arguments{
\item{x}{Path to the catch quantification files sorted by year or by vessel}

\item{spp_list}{A species list must be provided. The format is an R list of character vectors. The names must correspond to the ones used in BlackBox Analyzer Catch Quantification records.}
}
\value{
a dataset
}
\description{
Fix the catch quantification export format
}
