% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.CQ.R
\name{fix.CQ}
\alias{fix.CQ}
\title{Fix the catch quantification export format}
\usage{
fix.CQ(
  x =
    "Q:/scientific-projects/cctv-monitoring/data/blackbox extractions/catch_quantification/"
)
}
\arguments{
\item{x}{Path to the catch quantification files sorted by year or by vessel}
}
\value{
a dataset
}
\description{
Fix the catch quantification export format
}
