% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_logbooks.R
\name{logbook_import}
\alias{logbook_import}
\title{Format logbook / landings data to merge with EM data
Dataset preparations}
\usage{
logbook_import(x)
}
\arguments{
\item{x}{path to the directory where the logbook & sales notes are stored as .csv}
}
\value{
A dataset with all notes/annotations in long format, where rows are unique for hauls for no or one bycatch within that haul (each additional bycatch is listed as one supplementary row).
}
\description{
Format logbook / landings data to merge with EM data
Dataset preparations
}
